+++ engine/src/dskmac.cpp
@@ -2299,10 +2299,10 @@ static bool MCS_getentries_for_folder(MCStringRef p_folder, MCSystemListFolderEn
             t_entry.name = t_unicode_name;
             t_entry.data_size = t_catalog_infos[t_i] . dataLogicalSize;
             t_entry.resource_size = t_catalog_infos[t_i] . rsrcLogicalSize;
-            t_entry.creation_time = (uint32_t)t_creation_time;
-            t_entry.modification_time = (uint32_t) t_modification_time;
-            t_entry.access_time = (uint32_t) t_access_time;
-            t_entry.backup_time = (uint32_t) t_backup_time;
+            t_entry.creation_time = (uint64_t)t_creation_time;
+            t_entry.modification_time = (uint64_t)t_modification_time;
+            t_entry.access_time = (uint64_t)t_access_time;
+            t_entry.backup_time = (uint64_t)t_backup_time;
             t_entry.user_id = (uint32_t) t_permissions -> userID;
             t_entry.group_id = (uint32_t) t_permissions -> groupID;
             t_entry.permissions = (uint32_t) t_permissions->mode & 0777;
